Move MockFileManager to ApolloCodegenTestSupport target
