+++ Tests/ApolloCodegenTests/FileManagerExtensionsTests.swift
@@ -5,83 +5,6 @@ import ApolloCodegenLib
 import ApolloUtils
 import Nimble
 
-typealias FileAttributes = [FileAttributeKey : Any]
-
-/// A `FileManagerProvider` that can be used to mock the `apollo` extension as  found on `FileManager`.
-class MockFileManager: FileManagerProvider {
-  typealias fileExistsBlock = ((String, UnsafeMutablePointer<ObjCBool>?) -> Bool)
-  typealias removeItemBlock = ((String) throws -> ())
-  typealias createFileBlock = ((String, Data?, [FileAttributeKey : Any]?) -> Bool)
-  typealias createDirectoryBlock = ((String, Bool, [FileAttributeKey : Any]?) throws -> ())
-
-  enum BlocksCalled {
-    case fileExists
-    case removeItem
-    case createFile
-    case createDirectory
-  }
-
-  var fileExists: fileExistsBlock?
-  var removeItem: removeItemBlock?
-  var createFile: createFileBlock?
-  var createDirectory: createDirectoryBlock?
-  private(set) var blocksCalled: [BlocksCalled] = []
-
-  /// Initialize the instance with one or many of the method stubs to control the return values.
-  init(fileExists: fileExistsBlock? = nil,
-       removeItem: removeItemBlock? = nil,
-       createFile: createFileBlock? = nil,
-       createDirectory: createDirectoryBlock? = nil) {
-    self.fileExists = fileExists
-    self.removeItem = removeItem
-    self.createFile = createFile
-    self.createDirectory = createDirectory
-  }
-
-  func fileExists(atPath path: String, isDirectory: UnsafeMutablePointer<ObjCBool>?) -> Bool {
-    guard let block = fileExists else {
-      fail("fileExists block must be set before calling this method!")
-      return false
-    }
-
-    blocksCalled.append(.fileExists)
-    return block(path, isDirectory)
-  }
-
-  func removeItem(atPath path: String) throws {
-    guard let block = removeItem else {
-      fail("removeItem block must be set before calling this method!")
-      return
-    }
-
-    blocksCalled.append(.removeItem)
-    try block(path)
-  }
-
-  func createFile(atPath path: String, contents data: Data?, attributes attr: FileAttributes?) -> Bool {
-    guard let block = createFile else {
-      fail("createFile block must be set before calling this method!")
-      return false
-    }
-
-    blocksCalled.append(.createFile)
-    return block(path, data, attr)
-  }
-
-  func createDirectory(atPath path: String, withIntermediateDirectories createIntermediates: Bool, attributes: FileAttributes?) throws {
-    guard let block = createDirectory else {
-      fail("createDirectory block must be set before calling this method!")
-      return
-    }
-
-    blocksCalled.append(.createDirectory)
-    try block(path, createIntermediates, attributes)
-  }
-}
-
-/// Enables the `.apollo` etension namespace.
-extension MockFileManager: ApolloCompatible {}
-
 class FileManagerExtensionTests: XCTestCase {
   lazy var uniquePath: String = {
     CodegenTestHelper.outputFolderURL().appendingPathComponent(UUID().uuidString)
